requirements.txt
azure-functions
azure-functions-worker
azure-identity
azure-storage-blob
msgraph-sdk
openpyxl
ortools
pandas
xlsxwriter
